# spec/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Log Analyzer API
  version: 1.0.0
  summary: API for local-first log ingestion, analysis, incidents, and Ollama-powered insights.
  description: |
    This API is designed for a local backend that can be securely exposed over DuckDNS.
    It supports ingestion sources, parsed events, incident workflows, dashboard metrics,
    and AI analysis through Ollama with selectable model profiles.
servers:
  - url: https://your-subdomain.duckdns.org
    description: Public HTTPS endpoint behind reverse proxy
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/events:
    get:
      tags: [Events]
      summary: Query events
      parameters:
        - $ref: '#/components/parameters/FromParam'
        - $ref: '#/components/parameters/ToParam'
        - name: severity
          in: query
          schema:
            type: string
        - name: service
          in: query
          schema:
            type: string
        - name: host
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paged event results
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  next_cursor:
                    type: string
                    nullable: true
                required: [items]
  /api/v1/events/{id}:
    get:
      tags: [Events]
      summary: Get event details
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
...
  /api/v1/ai/analyze/window:
    post:
      tags: [AI]
      summary: Analyze a time window of events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                from:
                  type: string
                  format: date-time
                to:
                  type: string
                  format: date-time
                model_profile_id:
                  type: string
                  format: uuid
                max_events:
                  type: integer
                  minimum: 1
                  maximum: 5000
                  default: 500
              required: [from, to]
      responses:
        "202":
          description: Analysis queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncJobAccepted'
  /api/v1/ai/analyze/incident/{id}:
    post:
      tags: [AI]
      summary: Analyze one incident
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                model_profile_id:
                  type: string
                  format: uuid
      responses:
        "202":
          description: Analysis queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncJobAccepted'
  /api/v1/ai/chat:
    post:
      tags: [AI]
      summary: Ask a contextual question about selected logs/incidents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                model_profile_id:
                  type: string
                  format: uuid
                context:
                  type: object
                  properties:
                    incident_id:
                      type: string
                      format: uuid
                    event_ids:
                      type: array
                      items:
                        type: string
                        format: uuid
              required: [message]
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                  references:
                    type: array
                    items:
                      type: string
                required: [answer]
  /api/v1/ai/jobs/{id}:
    get:
      tags: [AI]
      summary: Get async AI job status and result
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: AI job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIJob'
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Exchange credentials for access token (optional local login mode)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required: [username, password]
      responses:
        "200":
          description: Access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    enum: [Bearer]
                  expires_in:
                    type: integer
                required: [access_token, token_type]
  /api/v1/auth/token:
    post:
      tags: [Auth]
      summary: Create API token with scoped permissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
                expires_at:
                  type: string
                  format: date-time
              required: [name, scopes]
      responses:
        "201":
          description: Created token (shown once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  token_id:
                    type: string
                    format: uuid
                required: [token, token_id]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FromParam:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date-time
    ToParam:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date-time
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        severity:
          type: string
        service:
          type: string
          nullable: true
        host:
          type: string
          nullable: true
        environment:
          type: string
          nullable: true
        event_type:
          type: string
          nullable: true
        message:
          type: string
        fields:
          type: object
          additionalProperties: true
        fingerprint:
          type: string
          nullable: true
      required: [id, source_id, timestamp, severity, message, fields]
...
    AIJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, completed, failed]
        result:
          type: object
          additionalProperties: true
          nullable: true
        error:
          type: string
          nullable: true
      required: [id, status]
    AsyncJobAccepted:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued]
      required: [job_id, status]
...